# docs/v1/swagger.json
{
    "swagger": "2.0",
    "info": {
...
        "title": "DFD API",
        "contact": {
            "name": "Seokho Song",
            "url": "http://github.com/devsdkGIN_MODE"
        },
        "license": {
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "devsdk.net/api/dfd",
    "basePath": "/",
    "paths": {
...
        "/v1/state": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create state\nPermission : **states.post**",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api/v1/state"
                ],
                "summary": "Create state",
                "parameters": [
                    {
                        "description": "Create user state",
                        "name": "state",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/docmodels.RequestBodyStatePost"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseBadRequest"
                        }
                    },
                    "401": {
                        "description": "Unauthorized Request. If token is expired, **token_expired** filed must be set true",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseUnauthorized"
                        }
                    },
                    "403": {
                        "description": "You don't have permission",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseNotFound"
                        }
                    },
                    "404": {
                        "description": "Cannt found user",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseInternalServerError"
                        }
                    }
                }
            }
        },
        "/v1/states": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get my state change history\nPermission : **states.get**",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api/v1/state"
                ],
                "summary": "Get my state history",
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/docmodels.ResponseSuccess"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "history": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.DFDHistory"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseBadRequest"
                        }
                    },
                    "401": {
                        "description": "Unauthorized Request. If token is expired, **token_expired** filed must be set true",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseUnauthorized"
                        }
                    },
                    "403": {
                        "description": "You don't have permission",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseNotFound"
                        }
                    },
                    "404": {
                        "description": "Cannt found user",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseInternalServerError"
                        }
                    }
                }
            }
        },
        "/v1/states/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get user's state change history\nPermission : **states.get**",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api/v1/state"
                ],
                "summary": "Get state history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/docmodels.ResponseSuccess"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "history": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.DFDHistory"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseBadRequest"
                        }
                    },
                    "401": {
                        "description": "Unauthorized Request. If token is expired, **token_expired** filed must be set true",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseUnauthorized"
                        }
                    },
                    "403": {
                        "description": "You don't have permission",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseNotFound"
                        }
                    },
                    "404": {
                        "description": "Cannt found user",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseInternalServerError"
                        }
                    }
                }
            }
        },
        "/v1/user": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get me\nPermission : **user.get**",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api/v1/user"
                ],
                "summary": "Get My User Information",
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/docmodels.ResponseSuccess"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "user": {
                                            "$ref": "#/definitions/models.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseBadRequest"
                        }
                    },
                    "401": {
                        "description": "Unauthorized Request. If token is expired, **token_expired** filed must be set true",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseUnauthorized"
                        }
                    },
                    "403": {
                        "description": "You don't have permission",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseNotFound"
                        }
                    },
                    "404": {
                        "description": "Cannt found user",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseInternalServerError"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "edit userfield.\nPermission : **user.patch**",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api/v1/user"
                ],
                "summary": "Edit user information",
                "parameters": [
                    {
                        "description": "username",
                        "name": "username",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/docmodels.RequestBodyPatchUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseBadRequest"
                        }
                    },
                    "401": {
                        "description": "Unauthorized Request. If token is expired, **token_expired** filed must be set true",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseUnauthorized"
                        }
                    },
                    "403": {
                        "description": "You don't have permission",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseNotFound"
                        }
                    },
                    "404": {
                        "description": "Cannt found user",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseInternalServerError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "docmodels.RequestBodyPatchUser": {
            "type": "object",
            "properties": {
                "profile_image_id": {
                    "type": "string",
                    "example": "6006d3cc95f8c8e32d660c04"
                },
                "username": {
                    "type": "string",
                    "example": "devsdk"
                }
            }
        },
        "docmodels.RequestBodyStatePost": {
            "type": "object",
            "properties": {
                "state": {
                    "type": "string",
                    "example": "Working"
                }
            }
        },
...
        "docmodels.ResponseBadRequest": {
            "type": "object",
            "properties": {
                "messsage": {
                    "type": "string",
                    "example": "Bad request message"
                },
                "status": {
                    "type": "integer",
                    "example": 400
                }
            }
        },
...
        "docmodels.ResponseInternalServerError": {
            "type": "object",
            "properties": {
                "messsage": {
                    "type": "string",
                    "example": "Internal Server Error"
                },
                "status": {
                    "type": "integer",
                    "example": 500
                }
            }
        },
...
        "docmodels.ResponseNotFound": {
            "type": "object",
            "properties": {
                "messsage": {
                    "type": "string",
                    "example": "Not found message"
                },
                "status": {
                    "type": "integer",
                    "example": 404
                }
            }
        },
        "docmodels.ResponseSuccess": {
            "type": "object",
            "properties": {
                "messsage": {
                    "type": "string",
                    "example": "success"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "docmodels.ResponseUnauthorized": {
            "type": "object",
            "properties": {
                "messsage": {
                    "type": "string",
                    "example": "Request Unauthorized message"
                },
                "status": {
                    "type": "integer",
                    "example": 401
                },
                "token_expired": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "models.DFDHistory": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "was": {
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "string",
                    "format": "date-time"
                },
                "discord_id": {
                    "type": "string"
                },
                "email": {
                    "type": "string",
                    "format": "email"
                },
                "id": {
                    "type": "string"
                },
                "lolUsername": {
                    "type": "string"
                },
                "modified": {
                    "type": "string",
                    "format": "date-time"
                },
                "profileImage": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}